+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Use a breadcrumb path to test only one particular part of the configuration node tree.
+
 ## v0.1.1
 
 - Fixed issue #1: ``ProcessedConfigurationEqualsConstraint`` had expected and actual value mixed up